code-files/requirements.txt
# Streamlit web framework
streamlit==1.31.0

# Azure OpenAI SDK (official Python SDK)
openai==1.12.0

# HTTP library for Azure AI Search REST API calls
requests==2.31.0

# Environment variable management
python-dotenv==1.0.1
